[Releng] Fix problems with the setup archiver
The latest ECF version no longer respects the time-out preferences and
uses a connection pool that limits the number of connections.  But I
think this pool implementation is buggy and the only way to work around
it is to increase the number of allowed collections.  At this point,
only the setup archiver needs to use this so I'm keeping the system
property hidden.