[Releng] Make Oomph.setup's org.eclipse.emf.ecoretools.sdk optional
In order to be able to add Oomph to the Platform SDK setup it would be
better if the requirement on org.eclipse.emf.ecoretools.sdk is optional
greedy so that failure to resolve this requirement doesn't cause
complete failure.  After all, one can work well on Oomph without the
diagram support.