[Releng] Make use of release train maker annotations
Extension models specify enablement annotations and these can provide
repositories that are effectively alternates for release train
repositories.  If a release train repository is present, it would be
better to omit the alternate repository so that requires are resolved
from the release train.  All the models are now properly marked to
indicate the repository is a release train alternate.

The setup task performer generates annotations on the synthesized
repository to record this information.

The product catalog generator marks repositories as release train
repositories.

The p2 director task, during consolidation determines if a release train
repository is present and removed any release train alternate
repositories in that case.

Overall this ensures that alternates are only used when needed, e.g.,
with the new Platform SDK product.