[549865] Installation of Eclipse fails if we chose to reject install unsigned components- Need a fix to remove settings which has been saved
While auto-accept is useful to remember and respect, auto-decline is
always questionable because of course the p2 director task will fail,
and  without a prompt, there is no recourse proceed with the
installation nevertheless.

The system properties SetupProperties.PROP_SETUP_UNSIGNED_POLICY and
SetupProperties.PROP_SETUP_CERTIFICATE_POLICY can be used to
auto-decline always if that is desired in general.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=549865