[Releng] Improve the installer's behavior
Keep track of the current highest LTS version of Java and, for
descriptor-based JRE, i.e., the JustJ JRES, select that version or
lower.  This is to avoid selecting Java 18 but rather Java 17 by default
when no local JREs are available.

Generate warnings about duplicate downloads only for pack.gz.  Often
this problem happens now because the shared connection times out but is
reused after timeout.

Make the simple dialog wider so that the JustJ JRE information is fully
visible.

Ensure that the "slow download" message doesn't show while the trust
dialog is showing.