Enable github validation for pull-requests
Currently there is no validation when open a PR for oomph code, this is
quite bad as it introduce the risk that after mering the code do not
build naymore or test fail without a contributor to get noticed about
that.

This adds the platform unified workflow for building maven based
projects.