Scan the bundle pool for missing files at the start of the p2 task
If the file is missing, the descriptor for it should be removed so that
it will be downloaded from scratch.

This should help prevent a recurring source of errors where it appears
that the virus scanner on Windows is deleting files from the pool
leaving it in a corrupted state.

https://github.com/eclipse-oomph/oomph/issues/60