Support `*.<filter-name>.filter.override.pattern` to induce filters
- Any key/value pair of this form in the SetupTaskContext is processed
such that the <filter-name> can be used as a filter where the value is
of the form `<java-pattern>-><replacement-string>` to induce
pattern/replace pairs that are matched against the value to be filtered
and, if the pattern matches, is replaced by the replacement using
Matcher.replaceAll. As such, the pattern's capture groups can be used in
the replacement.